+++ examples/html/place-react.jsx
@@ -1,4 +1,3 @@
-'use strict';
 
 function ShowCity(props) {
     return (